==> 4b - TestSim.Rmd <==
---
title: "Untitled"
output: pdf_document
---


# Setup

## Packages and Data

```{r}
library(tidyverse)

rm(list = ls())

source("0 - Functions - Simulation.R")

load("Data/Simulation Data/Sim Data.Rdata")

# ggplot(df.clusters, aes(UCS_Rank, SCS_Rank, color = factor(strata))) + 
#   geom_point() + 
#   facet_wrap(~ K) + 
#   theme_minimal()


```


## Specify Parameters and inputs

```{r}

sim.data <- df.sim %>% 
  select(DSID, all_of(covariates))

B.index.covariates <- c(covariates[!(covariates %in% c("Urban", "Suburban", "ToRu"))], "urbanicity")

B.index.formula <- as.formula(paste("accepted ~", paste(B.index.covariates, collapse = " + ")))

B.index.data <- df.sim %>%
  select(DSID, all_of(B.index.covariates))

df.K.ranks <- df.clusters %>%
              nest(cluster.ranks = -c(K)) %>%
              rename(K.condition = K)

params <- list(
  iterations = 12,
  RR.condition = unique(df.PS$RR),
  K.condition = c("K_05"),
  SB.condition = unique(df.PS$scale_factor)
)

# params <- list(
#   iterations = 2,
#   RR.condition = unique(df.PS$RR)[5],
#   K.condition = c("K_05"),
#   SB.condition = unique(df.PS$scale_factor)[4]
# )


param_df <- df.PS %>%
  nest(PS.data = -c(scale_factor, RR)) %>%
  rename(SB.condition = scale_factor,
         RR.condition = RR) %>%
  right_join(cross_df(params), 
             by = c("SB.condition", "RR.condition")) %>%
  left_join(df.K.ranks, 
            by = "K.condition") %>%
  mutate(B.index.formula = list(B.index.formula),
         B.index.data = list(B.index.data),
         sim.data = map(cluster.ranks, function(x) select(x, strata, DSID) %>% left_join(sim.data, by = "DSID")))



```

# Function Test

```{r}

test <- param_df %>% filter(SB.condition == 1 ,
                            RR.condition %in% c("RR_20", "RR_50"))

PS.data <- test$PS.data[[1]]
cluster.ranks <- test$cluster.ranks[[1]]
sim.data <- test$sim.data[[1]]

```

## Generate Potential Responses

```{r}


df.responses <- Generate_Responses(PS.data, cluster.ranks)

responses.multitest <- test %>%
  mutate(df.responses = pmap(.l = list(PS.data = PS.data, 
                               cluster.ranks = cluster.ranks), 
                     .f = Generate_Responses)) %>%
  select(SB.condition, RR.condition, K.condition, df.responses)




```

## Select Samples

```{r}
df.sampled <- Create_Samples(df.responses)


sampled.multitest <- responses.multitest %>%
  mutate(df.sampled = map(df.responses, Create_Samples)) %>%
  select(-df.responses)



```


## Calculate Statistics

### Recruitment Statistics
```{r}

df.recruitment.stats <- Calc_Recruitment_Stats(df.sampled) 

rstats.multitest <- sampled.multitest %>%
  mutate(df.recruitment.stats = map(df.sampled, Calc_Recruitment_Stats)) %>%
  select(-df.sampled)


```

### Track Sampled Schools
```{r}

df.samp.counts <- 
  df.sampled %>%
  filter(accepted) %>%
  select(sample_method, DSID) 


sampcounts.multitest <- 
  sampled.multitest %>%
  mutate(df.samp.counts = map(df.sampled, 
                              function(x) filter(x, accepted) %>% 
                                select(sample_method, DSID))) %>%
  select(-df.sampled)



```

### Means and SDs
```{r}
df.samp.stats <- 
  df.samp.counts %>%
  left_join(sim.data, 
            by = "DSID") %>%
  Calc_Sample_Statistics

sampstats.multitest <- 
  sampcounts.multitest %>%
  mutate(df.samp.stats = map(df.samp.counts, 
                             function(x) left_join(x, sim.data, by = "DSID") %>% 
                               Calc_Sample_Statistics)) %>%
  select(-df.samp.counts)


```

### B-Index
```{r}


df.B.indicies <- Bindex_Summary(df.sampled, B.index.data, B.index.formula)

bindex.multitest <- sampled.multitest %>%
  mutate(df.B.indicies = map(df.sampled, Bindex_Summary, B.index.data, B.index.formula)) %>%
  select(-df.sampled)


```


```{r}
test.results <- 
  test %>%
  mutate(df.responses = pmap(.l = list(PS.data = PS.data, 
                               cluster.ranks = cluster.ranks), 
                     .f = Generate_Responses),
         df.sampled = map(df.responses, 
                          Create_Samples),
         df.recruitment.stats = map(df.sampled, 
                                    Calc_Recruitment_Stats),
         df.samp.counts = map(df.sampled, 
                              function(x) filter(x, accepted) %>% 
                                select(sample_method, DSID)),
         df.samp.stats = map2(df.samp.counts, 
                              sim.data, 
                              function(x, y) left_join(x, y, by = "DSID") %>% 
                                Calc_Sample_Statistics),
         df.B.indicies = pmap(.l = list(df.sampled, 
                                        B.index.data, 
                                        B.index.formula), 
                              .f = Bindex_Summary)) %>%
  select(SB.condition, RR.condition, K.condition, df.recruitment.stats:df.B.indicies)



test.results

test.results$df.recruitment.stats[[1]]
test.results$df.samp.counts[[1]]
test.results$df.samp.stats[[1]]
test.results$df.B.indicies[[1]]
```


# Iteration Test

## Setup

### Packages and Data

```{r}
library(tidyverse)

rm(list = ls())

source("0 - Functions - Simulation.R")

load("Data/Simulation Data/Sim Data.Rdata")

# ggplot(df.clusters, aes(UCS_Rank, SCS_Rank, color = factor(strata))) + 
#   geom_point() + 
#   facet_wrap(~ K) + 
#   theme_minimal()


```


### Specify Parameters and inputs

```{r}

sim.data <- df.sim %>% 
  select(DSID, all_of(covariates))

B.index.covariates <- c(covariates[!(covariates %in% c("Urban", "Suburban", "ToRu"))], "urbanicity")

B.index.formula <- as.formula(paste("accepted ~", paste(B.index.covariates, collapse = " + ")))

B.index.data <- df.sim %>%
  select(DSID, all_of(B.index.covariates))

df.K.ranks <- df.clusters %>%
              nest(cluster.ranks = -c(K)) %>%
              rename(K.condition = K)

params <- list(
  iterations = 2,
  RR.condition = unique(df.PS$RR),
  K.condition = c("K_05"),
  SB.condition = unique(df.PS$scale_factor)
)


param_df <- df.PS %>%
  nest(PS.data = -c(scale_factor, RR)) %>%
  rename(SB.condition = scale_factor,
         RR.condition = RR) %>%
  right_join(cross_df(params), 
             by = c("SB.condition", "RR.condition")) %>%
  left_join(df.K.ranks, 
            by = "K.condition") %>%
  mutate(B.index.formula = list(B.index.formula),
         B.index.data = list(B.index.data),
         sim.data = map(cluster.ranks, function(x) select(x, strata, DSID) %>% left_join(sim.data, by = "DSID")))


test <- param_df %>% filter(SB.condition == 1 ,
                            RR.condition %in% c("RR_20", "RR_50"))

```

## Iteration Function

```{r}


Run_Iteration(x = 1,
              test$PS.data[[1]], 
              test$cluster.ranks[[1]], 
              test$sim.data[[1]], 
              test$B.index.data[[1]], 
              test$B.index.formula[[1]])

test %>%
  mutate(results = pmap(.l = list(x = 1, 
                                  PS.data, 
                                  cluster.ranks, 
                                  sim.data, 
                                  B.index.data, 
                                  B.index.formula),
                        .f = Run_Iteration))


# undebug(Run_Iteration)

```

## Iteration Timer

```{r}
Time_Iteration <- function(x = 1, 
                           PS.data, 
                           cluster.ranks, 
                           sim.data, 
                           B.index.data, 
                           B.index.formula) {
  system.time(Run_Iteration(x = x, 
                            PS.data, 
                            cluster.ranks, 
                            sim.data, 
                            B.index.data, 
                            B.index.formula))[3] / 60 
}
```


### Partial Conditions

```{r}


test.time <-
  Time_Iteration(x = 1,
                 test$PS.data[[1]], 
                 test$cluster.ranks[[1]], 
                 test$sim.data[[1]], 
                 test$B.index.data[[1]], 
                 test$B.index.formula[[1]])

system.time(multitest.time <- 
  test %>%
  mutate(minutes = pmap(.l = list(x = 1, PS.data, cluster.ranks, sim.data, B.index.data, B.index.formula),
                        .f = Time_Iteration)) %>%
  select(SB.condition, RR.condition, K.condition, minutes) %>% 
  unnest(minutes) %>%
  mutate(seconds = minutes * 60))

multitest.time
```

### All Conditions

```{r}

multitest.time <- 
  param_df %>%
  mutate(minutes = pmap(.l = list(x = 1, PS.data, cluster.ranks, sim.data, B.index.data, B.index.formula),
                        .f = Time_Iteration)) %>%
  select(SB.condition, RR.condition, K.condition, minutes) %>% 
  unnest(minutes) %>%
  mutate(seconds = minutes * 60)

multitest.time
```

# Parallel Test


## Setup

### Packages and Data

```{r}
library(tidyverse)
library(future)
library(furrr)

rm(list = ls())

source("0 - Functions - Simulation.R")

load("Data/Simulation Data/Sim Data.Rdata")

# ggplot(df.clusters, aes(UCS_Rank, SCS_Rank, color = factor(strata))) + 
#   geom_point() + 
#   facet_wrap(~ K) + 
#   theme_minimal()


```
### Specify Parameters and inputs

```{r}

sim.data <- df.sim %>% 
  select(DSID, all_of(covariates))

B.index.covariates <- c(covariates[!(covariates %in% c("Urban", "Suburban", "ToRu"))], "urbanicity")

B.index.formula <- as.formula(paste("accepted ~", paste(B.index.covariates, collapse = " + ")))

B.index.data <- df.sim %>%
  select(DSID, all_of(B.index.covariates))

df.K.ranks <- df.clusters %>%
              nest(cluster.ranks = -c(K)) %>%
              rename(K.condition = K)

params <- list(
  iterations = 2,
  RR.condition = unique(df.PS$RR),
  K.condition = c("K_05"),
  SB.condition = unique(df.PS$scale_factor)
)


param_df <- df.PS %>%
  nest(PS.data = -c(scale_factor, RR)) %>%
  rename(SB.condition = scale_factor,
         RR.condition = RR) %>%
  right_join(cross_df(params), 
             by = c("SB.condition", "RR.condition")) %>%
  left_join(df.K.ranks, 
            by = "K.condition") %>%
  mutate(B.index.formula = list(B.index.formula),
         B.index.data = list(B.index.data),
         sim.data = map(cluster.ranks, function(x) select(x, strata, DSID) %>% left_join(sim.data, by = "DSID")))


test <- param_df %>% filter(SB.condition == 1 ,
                            RR.condition %in% c("RR_20", "RR_50"))

```



## Test Driver

### map_dfr function

```{r}
multitest.results <- map_dfr(
  .x = 1:test$iterations[[1]],
  .f = Run_Iteration,
  PS.data = test$PS.data[[1]], 
  cluster.ranks = test$cluster.ranks[[1]], 
  sim.data = test$sim.data[[1]],
  B.index.data = test$B.index.data[[1]], 
  B.index.formula = test$B.index.formula[[1]]
)
```


```{r}

# 2 conditions, 2 iterations
test

# 1s condition, 2 iterations
multitest.results
```



### future_pmap function, 1 core

```{r}
Sim_Driver  <- function(iterations, 
                        PS.data, 
                        cluster.ranks, 
                        sim.data, 
                        B.index.data, B.index.formula) {
  

  results <- 
    map_dfr(
      .x = 1:iterations,
      .f = Run_Iteration,
      PS.data = PS.data, 
      cluster.ranks = cluster.ranks, 
      sim.data = sim.data,
      B.index.data = B.index.data, 
      B.index.formula = B.index.formula
      )
  
  results %>% return()
  
}

```


```{r}
plan(sequential)

multitest.results <- 
  test %>%
  mutate(results = future_pmap(.l = list(x = iterations,
                                         PS.data, 
                                         cluster.ranks, 
                                         sim.data, 
                                         B.index.data, 
                                         B.index.formula),
                             .f = Sim_Driver,
                             .options = furrr_options(seed = NULL)))


```


```{r}
# 2 conditions, 2 iterations
multitest.results

# first condition, 2 iterations
test.results <- multitest.results$results[[1]]

test.results$df.recruitment.stats[[1]] %>% count

```

### Timer

```{r}
plan(multisession(workers = 2))
system.time(multitest.results <- 
              test %>%
              mutate(minutes = future_pmap(.l = list(x = 1, 
                                                     PS.data, 
                                                     cluster.ranks, 
                                                     sim.data, 
                                                     B.index.data, 
                                                     B.index.formula),
                             .f = Time_Iteration,
                             .options = furrr_options(seed = NULL))) %>%
  select(SB.condition, RR.condition, K.condition, minutes) %>% 
  unnest(minutes) %>%
  mutate(seconds = minutes * 60))

plan(sequential)
```


```{r}
test$iterations <- 2

sim.test.results <- 
  system.time(multitest.results <- test %>%
    mutate(results = future_pmap(.l = list(iterations, PS.data, cluster.ranks, sim.data, B.index.data, B.index.formula),
                             .f = Sim_Driver,
                             .options = furrr_options(seed = NULL))) %>%
      select(SB.condition, RR.condition, K.condition, results))

(sim.test.results[[3]] / 2 * 1000) / 60 / 60

```

```{r}

param_df$iterations <- 2

sim.test.results <- 
  system.time(multitest.results <- param_df %>%
    mutate(results = future_pmap(.l = list(iterations, PS.data, cluster.ranks, sim.data, B.index.data, B.index.formula),
                             .f = Sim_Driver,
                             .options = furrr_options(seed = NULL))) %>%
      select(SB.condition, RR.condition, K.condition, results))

(sim.test.results[[3]] / 2 * 1000) / 60 / 60

multitest.results %>%
  unnest(results)
```

# Analysis Test


## Specify Conditions
```{r}

test <- param_df %>% filter(SB.condition == 1,
                            RR.condition %in% c("RR_10"))

PS.data <- test$PS.data[[1]]
cluster.ranks <- test$cluster.ranks[[1]]
sim.data <- test$sim.data[[1]]

```

## Generate Sample

```{r}

df.responses <- Generate_Responses(PS.data, cluster.ranks)

df.sampled <- Create_Samples(df.responses)

df.sampled %>%
  filter(sample_method == ("U_CS_UR_X2")) %>% 
  # filter(contacted) %>%
  ggplot(aes(x = UCS_Rank, y = PS, color = accepted)) +
  geom_point(alpha = .3, position = position_jitter()) +
  facet_grid(~contacted)

```

## Summarise Results
```{r}

df.recruitment.stats <- Calc_Recruitment_Stats(df.sampled) 

df.samp.counts <- df.sampled %>%
  filter(accepted) %>%
  select(sample_method, DSID) 

df.samp.stats <- df.samp.counts %>%
  left_join(sim.data) %>%
  Calc_Sample_Statistics

df.B.indicies <- Bindex_Summary(df.sampled, B.index.data, B.index.formula)

df.B.indicies

```


## Examine B-Index
```{r}
fit.predict <- function(x) tibble(PS = fitted(x), PS.logit = predict(x))


test.results <- df.sampled %>%
  select(sample_method, DSID, contacted, accepted) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, B.index.data),
         PS_sample = map(data, function(x) glm(data = x, formula = B.index.formula, family = quasibinomial()) %>% fit.predict()),
         accepted = map(data, select, "accepted")) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  group_by(sample_method) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted))
```

```{r}
test.results %>%
  unnest(data) %>%
  ggplot(aes(x = PS.logit)) +
  geom_density(alpha = .5, aes(fill = accepted)) +
  geom_text(data = test.results, aes(x = -9, y = 1, label = Bs)) +
  facet_wrap(~sample_method) +
  theme_minimal() +
  theme(legend.position = "none")

```

## Examine Alt B-Index
```{r}
B.index.covariates2 <- c(covariates[!(covariates %in% c("ToRu"))])
B.index.formula2 <- as.formula(paste("accepted ~", paste(B.index.covariates2, collapse = " + ")))


test.data <- df.sampled %>% 
  filter(sample_method == "S_BS_SR_X1") %>%
  select(sample_method, DSID, contacted, accepted) %>%
  left_join(df.sim)

test.data2 <- test.data %>%
  filter(accepted == T) %>%
  mutate(accepted = F) %>%
  bind_rows(test.data)

glm(data = test.data, formula = B.index.formula, family = quasibinomial) %>% 
  fit.predict %>% 
  bind_cols(test.data) %>%
  summarise(B = Calc_Bindex(PS.logit, accepted))

glm(data = test.data2, formula = B.index.formula, family = quasibinomial) %>% 
  fit.predict  %>% 
  bind_cols(test.data2) %>%
  summarise(B = Calc_Bindex(PS.logit, accepted))



```

```{r}

test.results <- df.sampled %>%
  select(sample_method, DSID, contacted, accepted) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_frm1 = map(data, function(x) glm(data = x, formula = B.index.formula, family = quasibinomial()) %>% fit.predict()),
         PS_frm2 = map(data, function(x) glm(data = x, formula = B.index.formula2, family = quasibinomial()) %>% fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  gather(key = PS_frm, value = PS_sample, PS_frm1, PS_frm2) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  group_by(sample_method, PS_frm) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))

test.results %>%
  filter(sample_method == "S_BS_SR_X1") %>%
  unnest(data) %>%
  select(-PS) %>%
  spread(PS_frm, PS.logit)

test.results %>%
  unnest(data) %>%
  ggplot(aes(x = PS.logit)) +
  geom_density(alpha = .5, aes(fill = accepted)) +
  geom_text(data = test.results, aes(x = -9, y = 1, label = Bs)) +
  facet_wrap(~sample_method) +
  theme_minimal() +
  theme(legend.position = "none")

```


# Best Case 

## UCS

```{r}

test <- param_df %>% filter(SB.condition == 1,
                            RR.condition %in% c("RR_50"))

PS.data <- test$PS.data[[1]]
cluster.ranks <- test$cluster.ranks[[1]]
sim.data <- test$sim.data[[1]]

df.responses <- Generate_Responses(PS.data, cluster.ranks)

df.sampled <- 
  df.responses %>%
  mutate(UCS_Rank = pmax(.001, UCS_Rank - .75)) %>%
  sample_frac(size = 1, weight = UCS_Rank ^ 10) %>%
  mutate(contacted = cumsum(Ej) <= 60,
         accepted = contacted  & Ej == 1,
         sample_method = "U_CS_UR_X2")


df.sampled %>%
  filter(contacted) %>%
  ggplot(aes(x = UCS_Rank, y = PS, color = accepted)) +
  geom_point(alpha = .3, position = position_jitter()) +
  facet_grid(~accepted)



df.sampled %>%
  filter(sample_method == "U_CS_UR_X2") %>%
  select(DSID, accepted, sample_method) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_sample = map(data, function(x) glm(data = x, 
                                               formula = B.index.formula, 
                                               family = quasibinomial()) %>% 
                           fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))

```

```{r}
cluster.ranks %>%
  left_join(df.PS %>% filter(scale_factor == 1, RR == "RR_50")) %>%
  arrange(desc(PS)) %>%
  ungroup() %>%
  mutate(accepted = (1:n() %in% c(1:60)),
         sample_method = "UCS")  %>%  
  ungroup() %>%
  select(DSID, accepted, sample_method) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_sample = map(data, function(x) glm(data = x, 
                                               formula = B.index.formula, 
                                               family = quasibinomial()) %>% 
                           fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))


```

```{r}
cluster.ranks %>%
  group_by(strata) %>%
  arrange(desc(SCS_Rank)) %>%
  mutate(accepted = (1:n() <= pa),
         sample_method = "SCS")  %>% 
  ungroup() %>%
  select(DSID, accepted, sample_method) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_sample = map(data, function(x) glm(data = x, 
                                               formula = B.index.formula, 
                                               family = quasibinomial()) %>% 
                           fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))

```

## Best Case SBS

```{r}
cluster.ranks %>%
  ungroup() %>%
  mutate(accepted = (SBS_Rank <= pa),
         sample_method = "SBS")  %>%
  select(DSID, accepted, sample_method) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_sample = map(data, function(x) glm(data = x, 
                                               formula = B.index.formula, 
                                               family = quasibinomial()) %>% 
                           fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))

```

```{r}
test.ranks <- df.sim %>%
  select(DSID, all_of(covariates)) %>%
  gather(key = variables, value = value, -DSID) %>%
  left_join(df.clusters %>% filter(K == "K_05") %>% select(DSID, strata)) %>%
  group_by(K, strata, variables) %>%
  mutate(w = 1) %>%
  mutate(value = w * (value - mean(value))^2) %>%
  # mutate(value = (value - mean(value))^2) %>%
  group_by(K, DSID, strata) %>%
  summarise(value = sqrt(sum(value))) %>%
  group_by(K, strata) %>%
  arrange(K, strata, value) %>%
  mutate(SBS_Rank = 1:n()) %>%
  select(-value)


test.ranks <- prop_allocations %>%
  select(K, strata, pa) %>%
  right_join(test.ranks) 

test.ranks %>%
  ungroup() %>%
  mutate(accepted = (SBS_Rank <= pa),
         sample_method = "SBS")  %>%
  select(DSID, accepted, sample_method) %>%
  nest(data = -sample_method) %>%
  mutate(data = map(data, left_join, df.sim),
         PS_sample = map(data, function(x) glm(data = x, 
                                               formula = B.index.formula, 
                                               family = quasibinomial()) %>% 
                           fit.predict()),
         accepted = map(data, select, c("accepted", "DSID"))) %>%
  select(-data) %>%
  unnest(cols = c(accepted, PS_sample)) %>%
  mutate(Bs = Calc_Bindex(PS, accepted) %>% round(2)) %>%
  nest(data = c(PS, PS.logit, accepted, DSID))

```
