==> 4c - PrepResults.Rmd <==
---
title: "Untitled"
output: pdf_document
---

# Setup

## Packages and Data

```{r}
library(papaja)
library(tidyverse)
library(kableExtra)
library(ggalt)
library(ggrepel)
library(ggforce)
library(concaveman)
library(cowplot)

rm(list = ls())


load("Data/Cluster Analysis/Clusters k2-10.rdata")
load("Data/Simulation Data/Sim Data.Rdata")

load(read_file("Data/Results/LastResults.txt"))



sim.results <- sim.results %>%
  transmute(SB = SB.condition,
            RR = parse_number(RR.condition),
            K = parse_number(K.condition),
            results = results) %>%
  unnest(results)

```


```{r}

ls() %>%
  tibble(Vars = .) %>%
  mutate(MB = map(Vars, function(x) object.size(eval(parse(text = x))) %>% format(units = "MB", digits = 3))) %>%
  unnest(MB) %>%
  arrange(desc(parse_number(MB)))

```

```{r}

df.list <- list()
results <- list()
vars <- list()

tabs <- list()
figs <- list()


vars$conditions <- c("SB", "RR", "Variables", "vnames") 
vars$methods <- c("Sampler", "Sample_Catagories", "Cluster", "Sample", "Method_Label") 
vars$cluster <- list(mixed = cluster_vars, dummy = covariates)
```

```{r}
df.list$strata.id <- 
  df.clusters %>%
  ungroup() %>%
  filter(K == "K_05") %>%
  mutate(strata = factor(paste("Strata", strata))) %>%
  select(DSID, strata, SBS_Rank, UCS_Rank, SCS_Rank)

```

```{r}

df.list$pop.data <- 
  left_join(df.sim, df.list$strata.id, by = "DSID")


```

```{r}


df.list$PS <- df.PS %>% 
  mutate(logit = log(PS / (1 - PS)),
         RR = parse_number(RR))


```

```{r}
df.list$coefs <- DGM.Bs %>%
  transmute(vnames = var, log_odds = Bs) %>%
  full_join(read_csv("Data/Paper/Clustering Variables.csv",
                      col_types = cols()),
            by = "vnames") %>%
  full_join(intercepts, by = character()) %>%
  mutate(Coeficient = log_odds * scale_factor,
         RR = parse_number(RR)) %>%
  rename(SB = scale_factor,
         Intercept = Int) %>%
  select(Sub, Category, Type, vnames, Variables, SB, RR, log_odds, Coeficient, Intercept) %>%
  nest(DGM = c(SB, RR, Coeficient, Intercept))

```


```{r}
df.list$pop.stats <- df.list$pop.data %>%
  select(DSID, all_of(covariates), strata) %>%
  gather(key = vnames, value = value, all_of(covariates)) %>%
  group_by(vnames) %>%
  mutate(T.SS = var(value) * (n() - 1)) %>%
  group_by(vnames, T.SS, strata) %>%
  summarise(W.SS = var(value) * (n() - 1), 
            .groups = "drop_last") %>%
  summarise(W.SS = sum(W.SS), 
            .groups = "drop_last") %>%
  mutate(B.SS = T.SS - W.SS,
         vratio = B.SS / T.SS) %>%
  left_join(df.list$coefs, 
            by = "vnames") %>%
  left_join(df.pop.stats %>% rename(vnames = var), 
            by = "vnames")

df.list$k.stats <- cluster.stats

df.list$cluster.pa <- df.clusters %>%
  group_by(K, strata, pa) %>%
  count()

```


```{r}
rename_results <- function(result.data) {
  result.data %>%
    separate(sample_method, into = c("Cluster", "Sample", "Rank_Type", "Rank_Factor"), remove = F, fill = "right") %>%
    rename(Sampler = sample_method) %>%
    mutate(Method_Label = paste(Cluster, Sample, sep = ""),
           Cluster = factor(Cluster, 
                            level = c("S", "U"), 
                            label = c("Stratified", "Unstratified")),
           Sample = factor(Sample,
                           level = c("BS", "CS", "RS"), 
                           label = c("Balanced", "Convenience", "Random")),
           Rank_Type = factor(Rank_Type,
                              level = c("SR", "UR"), 
                              label = c("Stratified Ranks", "Unstratified Ranks")),
           Rank_Factor = factor(Rank_Factor)) %>%
    pack(Sample_Catagories = c(Cluster, Sample, Rank_Type, Rank_Factor))
}



results$B <- 
  sim.results %>% 
  select(SB:K, B) %>%
  unnest(B) %>%
  rename_results


# results$B  %>%
#   # filter(SB == 1, RR == 10, Sampler == "U_CS_UR_X2") %>%
#   arrange(Bs)

# results$B %>% filter(is.na(Bs))

results$r.stats <- sim.results %>% 
  select(SB:K, r.stats) %>%
  unnest(r.stats) %>%
  rename_results 


results$smd.stats <- sim.results %>% 
  select(SB:K, smd.stats) %>%
  unnest(smd.stats) %>% 
  rename(vnames = var) %>% 
  rename_results() %>%
  left_join(df.list$pop.stats %>% unnest(DGM), by = c("SB", "RR", "vnames")) %>%
  mutate(mSMD = (sim.mean - pop.mean) / pop.sd,
         abs.mSMD = abs(mSMD))




```


```{r}


calc_gini_curve <- function(count.data, pop.list) {
  count.data %>%
    full_join(pop.list, by = "DSID") %>%
    mutate(count = ifelse(is.na(n), 0, n),
           prop.selected = count / sum(count),
           n = n(),
           prop.pop = 1/n) %>%
    arrange(prop.selected) %>%
    transmute(c.prop.sel = cumsum(prop.selected),
              c.prop.pop = cumsum(prop.pop),
              prop.selected = prop.selected) %>%
    as.data.frame
}



results$gini <- sim.results %>% 
  select(SB:K, samp.counts) %>%
  unnest(samp.counts) %>%
  nest(count.data = c(DSID, n)) %>%
  rename_results() %>%
  mutate(count.data = map(count.data, calc_gini_curve, select(df.sim, DSID)),
         Gini_Coef = map_dbl(count.data, function(x) ineq::Gini(x$prop.selected)))

# results$gini %>% unpack(Sample_Catagories)
```


```{r}
rm(list = c("cluster.stats", 
            "df.clusters", 
            "df.pop.stats", 
            "df.PS",
            "rename_results",
            "DGM.Bs",
            "sim.results",
            "prop_allocations",
            "clust_time",
            "calc_gini_curve",
            "covariates",
            "K",
            "minK",
            "cluster_vars"))


```

```{r}
ls() %>%
  tibble(Vars = .) %>%
  mutate(MB = map(Vars, function(x) object.size(eval(parse(text = x))) %>% format(units = "MB", digits = 3))) %>%
  unnest(MB) %>%
  arrange(desc(parse_number(MB)))


save.image(file = "Data/Paper/Summarised Sim Results.RData")


```
