==> 5 - Analysis.Rmd <==
---
title             : "GenSamp: RESULTS"
shorttitle        : "RESULTS"

author:
  - name          : "Gleb Furman"
    affiliation   : "1"
    corresponding : yes    # Define only one corresponding author
    address       : "University of Texas at Austin, SZB 504, 1912 Speedway, Austin, Texas 78712"
    email         : "gleb.furman@utexas.edu"
  - name          : "James E. Pustejovsky"
    affiliation   : "2"
  - name          : "Elizabeth Tipton"
    affiliation   : "3"
    

affiliation:
  - id            : "1"
    institution   : "University of Texas at Austin"
  - id            : "2"
    institution   : "University of Wisconsin-Madison"
  - id            : "3"
    institution   : "Northwestern University"


abstract: |
    <!-- In order for educational research to be informative to policy makers, studies must be designed to make robust estimates of causal effects at the population level. Large scale multi-site randomized trials (MRT) often rely on vague convenience sampling methodology when recruiting districts and schools, resulting in relatively homogeneous samples that may differ greatly from the intended population of interest. Retrospective methods that quantify and statistically adjust for those differences are promising but have limited effect when the sample differs greatly from the population. Designing sampling methods that focus on generalizability may be a more effective but costly solution, but limited methodological research has been performed to examine their effectiveness in the educational context. This paper examines one promising method, stratified balanced sampling (SBS), in the context of recruiting a representative sample of schools for a large scale MRT. Using simulations based on real data, we compare SBS to stratified and unstratified versions of convenience sampling and probability sampling. Several models for generating school participation and emulating convenience sampling are proposed. Results indicate that SBS and stratified random sampling (SRS) result in highly generalizable samples. These methods are extremely costly to implement, however, especially when the population average willingness to participate is low. Stratified convenience sampling (SCS) is a potential compromise. -->

# keywords          : "keywords"
# wordcount         : "X"

bibliography      : ["references.bib"]

figsintext        : yes
figurelist        : no
tablelist         : no
footnotelist      : no
lineno            : no
mask              : no

documentclass     : "apa6"
classoption       : "man"
output            : papaja::apa6_pdf
---


```{r setup, include = FALSE}
rm(list = ls())

output.screen <- T

knitr::opts_chunk$set(echo = F,
                      cache = F,
                      message = output.screen,
                      warning = output.screen,
                      results = "asis",
                      knitr.table.format = "latex")
options(knitr.kable.NA = "-")

```

# Setup

## Packages and Data

```{r load.packages, include = FALSE}
library(papaja)
library(tidyverse)
library(kableExtra)
library(ggalt)
library(ggrepel)
library(ggforce)
library(concaveman)
library(cowplot)

rm(list = ls())

```

```{r}
load(file = "Data/Paper/Summarised Sim Results.RData")

source("0 - Functions - Analysis and Paper.R")

```


## Organize Objects

```{r}

grey.plots <- F


lab.vars <- list(k = "Number of strata (k)", 
                 vratio = expression(paste("Variance Ratio (",
                                           sigma[B]^2/sigma[T]^2,
                                           ")")), 
                 ch = expression("CH index (" ~ 10^3 ~ ")"),
                 n_j = "Target sample per cluster",
                 pop.prop = "Proportion of population (%)",
                 smd = "Standardized mean difference",
                 vars = "Participation covariates",
                 intercept = "Participation model intercept",
                 rr = "Potential participation rate (%)",
                 B = "Average B-index",
                 coef = "Participation model coeficients \n(absolute log-odds)",
                 contacts = "Schools contacted ('000s)",
                 rr.o = "Population participation  \n(observed %)")
```

# Data Summary


## Covariate Statistics

```{r}
tabs$pop.desc <- 
  df.list$pop.stats %>%
  ungroup() %>%
  arrange(Category, Sub) %>%
  select(Category, Variables, pop.mean, pop.sd, log_odds, vratio) %>%
  mutate(
    M = round(pop.mean, 2),
    SD = round(pop.sd, 2),
    .keep = "unused",
    SF100 = (log_odds * 10) %>% round(3) / 10,
    vratio = round(vratio, 3)
  )

```

## Continuous variable distributions

```{r}
tabs$dist <- 
  df.list$pop.data %>%
  select(all_of(vars$cluster$mixed)) %>%
  select_if(is.numeric) %>%
  mutate(n = log(n),
         ST.ratio = log(ST.ratio),
         dSCH = log(dSCH)) %>%
  gather(key = vnames, value = val) %>%
  left_join(df.list$coefs,
            by = "vnames")

figs$dist <- 
  tabs$dist %>%
  ggplot(aes(x = val)) +
  geom_histogram() +
  scale_y_continuous(labels = function(x) x / 100) +
  facet_wrap(~ Variables, scales = "free", ncol = 3) +
  labs(x = "", y = "Schools (100's)")

figs$dist

```


# Methods Summary

## Cluster Analysis

### Selecting k
```{r}
hlines <- 
  df.list$k.stats %>%
  filter(K %in% c(2, 4)) %>%
  select(ch) %>%
  unlist

figs$ch <- 
  df.list$k.stats %>%
  ggplot(aes(x = K, y = ch)) +
  geom_point() +
  geom_line() +
  geom_hline(yintercept = hlines, linetype = "dotted") +
  scale_x_continuous(breaks = c(2:10)) +
  scale_y_continuous(breaks = seq(5000, 8500, 1000),
                     limits = c(5000, 8500),
                     labels = function(x) x / 1000) +
  labs(y = lab.vars$ch,
       x = lab.vars$k)

figs$ch
```

```{r}



figs$vratio <- 
  df.list$k.stats %>%
  bind_rows(summarise(., K = 1, bt.ss = 0)) %>%
  mutate(min80 = sum(bt.ss <= .8) + min(K) - .5) %>%
  ggplot(aes(x = K, y = bt.ss)) +
  geom_point() +
  geom_vline(aes(xintercept = min80), linetype = "dotted") +
  geom_hline(yintercept = .8, linetype = "dotted") +
  scale_x_continuous(breaks = 1:10) +
  scale_y_continuous(breaks = seq(0, 1, .25), limits = c(0, 1)) +
  geom_line() +
  labs(y = lab.vars$vratio,
       x = lab.vars$k)

figs$vratio
```

```{r}




tabs$allocation <- 
  df.list$cluster.pa %>%
  mutate(k = parse_number(K),
         strata = paste("Strata", strata))

figs$allocation <- 
  tabs$allocation %>%
  ggplot(aes(x = k, y = pa)) +
  geom_point(shape = "-", size = 15, alpha = .5) +
  labs(y = lab.vars$n_j,
       x = lab.vars$k) +
  scale_x_continuous(breaks = seq(2, 10, 1)) +
  scale_y_continuous(breaks = seq(0, 50, 5),
                     limits = c(0, 35),
                     sec.axis = sec_axis(~./.6, 
                                         name = lab.vars$pop.prop,
                                         breaks = seq(0, 100, 10))) +
  stat_function(fun = function(x) 60/x, geom = "line", linetype = "dashed", size = .5) +
  geom_hline(yintercept = 5, linetype = "dotted", size = .5)

figs$allocation

```

```{r}

figs$joint.cluster.stats <- 
  plot_grid(plotlist = list(figs$ch,
                            figs$vratio), 
          labels=c("(a)", "(b)"), 
          ncol = 2, 
          nrow = 1, 
          rel_heights = c(1, 1),
          label_x = .5,
          align = "v",
          label_fontface = "plain" ,
          label_size = 10) %>%
  plot_grid(figs$allocation, 
            ncol = 1,
            nrow = 2, 
            labels = c("", "(c)"),
            label_x = .5,
            # align = "h",
            label_fontface = "plain" ,
            label_size = 10)

figs$joint.cluster.stats 
```

```{r, results = "asis"}

tabs$cluster.SMD <- 
  df.list$pop.data %>% 
  select(strata, all_of(vars$cluster$dummy)) %>%
  mutate(strata = str_replace(strata, "Strata", "Stratum")) %>%
  gather(key = vnames, value = val, -strata) %>%
  group_by(strata, vnames) %>%
  summarise(strat.mean = mean(val), .groups = "drop_last") %>%
  left_join(df.list$pop.stats, by = "vnames") %>%
  transmute(vnames = vnames,
            SMD = round((pop.mean - strat.mean) / pop.sd, 2)) %>%
  left_join(df.list$coefs, by = "vnames")

```


```{r}


figs$cluster.SMD <- 
  tabs$cluster.SMD %>%
  mutate(sig = ifelse(abs(SMD) > .25, "Sig", "Not Sig"),
         neg = SMD > 0,
         k = parse_number(as.character(strata))) %>%
  ggplot(aes(x = SMD, y = Variables, color = strata, alpha = sig, shape = strata)) +
  geom_point() +
  geom_vline(xintercept = c(-.25, .25), linetype = "dotted") +
  scale_alpha_discrete(guide = "none")  +
  scale_x_continuous(limits = c(-2,2)) +
  # theme(axis.text.y = element_blank(),
  #       axis.ticks.y = element_blank()) +
  labs(x = lab.vars$smd ,
       y = lab.vars$vars)

figs$cluster.SMD
```





## Variation explained by the strata
```{r}

tabs$strata.data <- 
  tabs$allocation %>%
  filter(k == 5) %>%
  left_join(df.list$k.stats %>% filter(K == 5) %>%
              rename(k = K), by = "k") %>%
  mutate(bet.perc = round(bt.ss * 100, 2),
         ch = round(ch, 2))




```

## Participation Generating Model

```{r}

 
df.list$coefs %>% 
  select(Sub, Category, Type, Variables, log_odds) %>%
  unique() %>%
  kable() %>%
  kable_styling()


```


```{r}


tabs$intercepts <- 
  intercepts %>%
  mutate(RR = parse_number(RR),
         scale_factor = factor(scale_factor),
         Intercept = Int)

tabs$intercepts %>%
  ggplot(aes(x = RR, y = Intercept, group = scale_factor, color = scale_factor)) +
  geom_point() +
  geom_line() +
  scale_x_continuous(breaks = seq(0, 100, 10)) +
  labs(y = lab.vars$intercept,
       x = lab.vars$rr)
```




```{r, fig.cap = "Distributions of Participation Logits"}

df.list$PS %>%
  mutate(RR = as_factor(RR)) %>%
  ggplot(aes(x = logit, y = RR, fill = RR)) +
  ggridges::geom_density_ridges(stat = "binline", 
                                bins = 100, 
                                alpha = .3, 
                                scale = 2) +
  theme_apa() +
  theme(legend.position = "none")+
  scale_x_continuous(breaks = seq(-8,10, 2))  +
  facet_wrap(~scale_factor)
```

# Results

## Generalizability

### B Index

```{r}

get_quant <- function(x) {
  tibble(
    Q = quantile(x), 
    Percent = paste("q_", 
                    parse_number(names(Q)), 
                    sep = "")) %>% 
    spread(Percent, Q)
}

set_compare <- function(data, variable = "Sampler", 
                        value = "S_BS_SR_X1", 
                        observations = c("M", "q_25", "q_50", "q_75")){
  data %>%
  filter(.[[variable]] == value) %>%
  nest(Comparison = c(variable, observations)) %>%
  select(SB:K,
         Comparison) %>%
  mutate(Comparison = map(Comparison, rename_with, function(x) paste(x, "_C", sep = ""))) %>%
  unnest(Comparison)
}

tabs$B <- 
  results$B %>%
  chop(Bs) %>%
  mutate(M = map_dbl(Bs, mean),
         SD = map_dbl(Bs, sd),
         LL = M - SD,
         UL = M + SD,
         Quantile = map(Bs, get_quant)) %>%
  select(-Bs) %>%
  unpack(Sample_Catagories) %>%
  unnest(Quantile) %>%
  mutate(SB.factor = paste("SF = ", SB, sep = ""))

tabs$B.labs <-  
  tabs$B %>%
  filter(
    !(Rank_Factor %in% c("X9")),
    SB > 0.01,
    RR == 10,
  ) 

seeds <- list()
# seeds$fig.B <- floor(runif(1) * 10^4)
seeds$fig.B <- 1841
  
```

```{r, fig.cap = "Averge B-Index"}

figs$B <-
  tabs$B %>%
  filter(
    !(Rank_Factor %in% c("X9")),
    SB > 0.01
  ) %>%
  ggplot(aes(x = RR, y = M)) %>%
  common_line() %>%
  common_point() +
  # theme_minimal() +
  labs(y = lab.vars$B,
       x = lab.vars$rr) +
  annotate(geom = "rect",
           ymax = .95,
           ymin = -Inf,
           xmax = -Inf,
           xmin = Inf,
           fill = "white",
           alpha = .4) + 
  annotate(geom = "rect",
           ymax = .8,
           ymin = -Inf,
           xmax = -Inf,
           xmin = Inf,
           fill = "white",
           alpha = .4) +
  geom_hline(yintercept = c(.8, .95), linetype = "dotted", alpha = .5)  +
  # facet_grid(SB.factor ~.) +
  facet_grid(~ SB.factor)  +
  scale_x_continuous(breaks = seq(0, 100, 20), minor_breaks = NULL, limits = c(-5, 100)) +
  scale_y_continuous(breaks = seq(.6, 1, .1), limits = c(.6, 1)) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$B.labs,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = -15,
                   seed = seeds$fig.B,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 1) +
  theme_minimal() +
  theme(legend.position = "none")


figs$B

```




```{r}
tabs$B %>%
  left_join(set_compare(.), by = c("SB", "RR", "K")) %>%
  filter(SB == 1,
         Sampler != Sampler_C,
         !(Rank_Factor %in% c("X9"))) %>%
  ggplot(aes(x = RR, y = M, group = Sampler, color = Cluster, fill = Cluster)) +
  geom_line() +
  geom_ribbon(aes(ymin = q_25, ymax = q_75), alpha = .1, linetype = "dotted") +
  geom_line(aes(y = M_C), color = "black", alpha = .2) +
  geom_ribbon(aes(ymin = q_25_C, ymax = q_75_C), alpha = .1, linetype = "dotted", color = "black") +
  geom_hline(yintercept = c(.8, .95), linetype = "dotted", alpha = .5) +
  facet_wrap(~ Sample + Rank_Factor, nrow = 1) +
  theme_minimal() +
  theme(legend.position = "bottom") +
  labs(y = lab.vars$B,
       x = lab.vars$rr)
  

```




### SMD

```{r, fig.cap = "Average SMDs", fig.height = 7}

tabs$smd.results <- 
  results$smd.stats %>%
  select(any_of(vars$methods), any_of(vars$conditions),
         strata, log_odds, mSMD, abs.mSMD) %>%
  unpack(Sample_Catagories) %>%
  filter(
    !(Rank_Factor %in% c("X9")),
    strata == 0,
    SB > 0.01
  ) %>%
  select(-strata, -log_odds) %>%
  mutate(
    Value = abs.mSMD,
    alpha.group = as.numeric(Value > .25),
    SF_fac = paste("SF =", SB)
  )

tabs$smd.by.scale <- 
  tabs$smd.results %>%
  nest(smd.plot.data = -c(Variables, vnames)) %>%
  mutate(plots = pmap(.l = list(data = smd.plot.data, 
                                g.title = Variables), 
                      .f = plot_smd, 
                      l.pos = "bottom"),
         plots = setNames(plots, vnames))

figs$smd.by.scale <-
  tabs$smd.by.scale$plots %>% 
  as.list




```



```{r}
# Group 1
figs$smd.by.scale$pFem
figs$smd.by.scale$ST.ratio

```

```{r}
# Group 2
figs$smd.by.scale$Suburban
figs$smd.by.scale$T1
figs$smd.by.scale$Urban


```

```{r}
# Group 3
figs$smd.by.scale$pTotfrl
figs$smd.by.scale$ethHisp
figs$smd.by.scale$ethWhite
figs$smd.by.scale$pELL
figs$smd.by.scale$dSCH
figs$smd.by.scale$ToRu
```


```{r}
# Group 4
figs$smd.by.scale$n
figs$smd.by.scale$ethBlack

```



### SMD Groups

```{r}
vars$smd.groups <- 
  tibble(`Group 1` = list(c("pFem", "ST.ratio")),
         `Group 2` = list(c("Suburban", "T1", "Urban")),
         `Group 3` = list(c("pTotfrl", "ethHisp", "ethWhite", "pELL", "dSCH", "ToRu")),
         `Group 4` = list(c("n", "ethBlack"))) %>%
  pivot_longer(cols = everything(), names_to = "Group", values_to = "vnames") %>%
  unnest(vnames)

```

#### Example all SF

```{r}
vars$smd.examples <-
  vars$smd.groups %>%
  filter(vnames %in% c("Suburban", "dSCH", "n", "ST.ratio"))

tabs$smd.examples <- 
  vars$smd.examples %>%
  left_join(tabs$smd.results, by = "vnames")

figs$smd.examples <- 
  tabs$smd.examples %>% 
  ggplot(aes(x = RR, y = Value)) %>%
  common_line() %>%
  common_point()  + 
  annotate(geom = "rect",
           ymax = .25,
           ymin = -Inf,
           xmax = -Inf,
           xmin = Inf,
           fill = "white",
           alpha = .5) +
  labs(y = lab.vars$smd,
       x = lab.vars$rr) + 
  scale_x_continuous(breaks = seq(10, 100, 30),
                     limits = c(10, 100)) +
  facet_grid(Group ~ SB, scales = "free_y") +
  geom_hline(yintercept = c(.25), linetype = "dotted") +
  geom_hline(yintercept = 0, linetype = "solid")

figs$smd.examples 

```


#### Example SF = 1

```{r}


tabs$smd.sf100.labs <-  
  tibble(Method_Label = c("SBS", "URS", "UCS", "SRS", "SCS"),
       RR = c(10, 30, 50, 70, 90),
       SB = 1) %>%
  left_join(tabs$smd.examples, 
            by = c("Method_Label", "RR", "SB")) %>%
  mutate(Group = paste(Group, 
                       Variables, 
                       sep = ":\n"))


seeds <- list()
# seeds$fig.B <- floor(runif(1) * 10^4)
seeds$fig.smd.sf100.labs <- 1841

figs$smd.examples.sf100 <-
  tabs$smd.examples %>%
  filter(SB == 1) %>%
  mutate(Group = paste(Group, 
                       Variables, 
                       sep = ":\n")) %>%
  ggplot(aes(x = RR, y = Value)) %>%
  common_line() %>%
  common_point() +
  # facet_wrap(~ Group, scales = "free_y") +
  facet_wrap(~ Group) +
  geom_hline(yintercept = c(.25), linetype = "dotted") +
  geom_hline(yintercept = 0, linetype = "solid") +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$smd.sf100.labs,
                   direction = "y",
                   # force_pull = 0,
                   # force = .5,
                   nudge_y = .50,
                   nudge_x = 5,
                   seed = seeds$fig.smd.sf100.labs,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 0) + 
  annotate(geom = "rect",
           ymax = .25,
           ymin = -Inf,
           xmax = -Inf,
           xmin = Inf,
           fill = "white",
           alpha = .4) +
  labs(y = lab.vars$smd,
       x = lab.vars$rr) 

figs$smd.examples.sf100
```

```{r}

figs$smd.examples.sf100.v1 <-
  figs$smd.examples.sf100 +
  scale_x_continuous(breaks = seq(0, 100, 20), minor_breaks = NULL, limits = c(-5, 100)) +
  scale_y_continuous(breaks = seq(0, 1.75, .5), limits = c(-.25, 1.6)) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$smd.sf100.labs %>% filter(Group != "Group 1"),
                   direction = "y",
                   force_pull = 0,
                   force = 1,
                   nudge_x = -7,
                   seed = seeds$fig.smd.sf100.labs,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 2) +
  facet_wrap(~ Group) +
  theme_minimal() +
  theme(legend.position = "none")

figs$smd.examples.sf100.v1

```

```{r}
smd.limits <- 
  function(x)  {
    newlim <- c(-.1, ceiling(max(x) * 10) / 10)
    newlim
    }

smd.break <- 
  function(x) {
    x.max <- max(smd.limits(x))
    x.break <- ceiling(x.max * 2) / 10
    newbreak <- seq(0, x.max, x.break)
    newbreak
}

figs$smd.examples.sf100.v2 <-
  figs$smd.examples.sf100 +
  facet_wrap(~ Group, scales = "free_y") +
  scale_x_continuous(breaks = seq(0, 100, 20), minor_breaks = NULL, limits = c(-5, 100)) +
  # scale_y_continuous(breaks = seq(0, 1.75, .25)) +
  scale_y_continuous(limits = smd.limits, breaks = smd.break) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$smd.sf100.labs,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = -10,
                   seed = seeds$fig.smd.sf100.labs,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 0) +
  facet_wrap(~ Group, scales = "free_y") +
  theme_minimal() +
  theme(legend.position = "none")


figs$smd.examples.sf100.v2
```



### Auto-Grouping

```{r}
# note: all smd <= .25 considered 0

test <- 
  tabs$smd.results %>%
  select(comp_Method = Method_Label,
         comp_Cluster = Cluster,
         comp_Sample = Sample,
         SB, 
         RR, 
         vnames, 
         comp = abs.mSMD) %>%
  full_join(tabs$smd.results, by = c("SB", "RR", "vnames")) %>%
  filter(Method_Label != comp_Method) %>%
  select(SB,
         RR, 
         Variables, 
         vnames, 
         Method = Method_Label, 
         comp_Method, 
         abs.mSMD, 
         comp,
         Cluster,
         Sample,
         comp_Cluster,
         comp_Sample) %>%
  arrange(vnames, SB, RR, Method) 

test.group <- 
  test %>%
  mutate(below_25 = abs.mSMD <= .25,
         abs.mSMD = ifelse(abs.mSMD <= .25, 0, abs.mSMD),
         comp = ifelse(comp <= .25, 0, comp),
         less = abs.mSMD <= comp,
         greater = abs.mSMD > comp) %>%
  group_by(SB, Variables, vnames, across(contains(c("Method", "Cluster", "Sample")))) %>%
  summarise(below_25 = prod(below_25),
            less = prod(less),
            greater = prod(greater),
            .groups = "drop") 

```


#### Group 1

```{r}
# Group 1
# All methods perform below .25

test.g1 <- 
  test.group %>%
  select(SB, vnames, Method, below_25) %>% 
  unique() %>%
  arrange(SB, vnames) %>%
  # filter(SB == 1) %>%
  group_by(SB, vnames) %>%
  summarize(g1_a = prod(below_25),
            .groups = "drop_last") %>%
  filter(g1_a == 1) %>%
  mutate(Group = "Group 1") %>%
  select(SB, vnames, Group) %>%
  arrange(vnames, SB) 
  

test.g1
```

```{r}
figs$smd.by.scale[unique(test.g1$vnames)]

```

#### Group 2
```{r}
# Group 2
# a = smd below .25 across RR
# b = stratified methods
# c = unstratified methods
# ab = all stratified methods with smd below .25 across RR
# ac = no unstratified methods with smd below .25 across RR


test.g2 <-
  test.group %>%
  select(SB, vnames, Method, Cluster, below_25) %>% 
  unique() %>%
  arrange(SB, vnames) %>%
  # filter(SB == 1) %>%
  group_by(SB, vnames, Cluster) %>%
  summarize(g2_a = prod(below_25),
            g2_b = prod(Cluster == "Stratified"),
            g2_c = 1 - g2_b,
            .groups = "drop_last") %>%
  summarise(g2_ab = sum(g2_a * g2_b),
            g2_ac = 1 - sum(g2_a * g2_c),
            .groups = "drop") %>%
  filter(g2_ab * g2_ac == 1) %>%
  mutate(Group = "Group 2") %>%
  select(SB, vnames, Group) %>%
  arrange(vnames, SB) 
  
test.g2 
```

```{r}
figs$smd.by.scale[unique(test.g2$vnames)]
```


#### Group 3

```{r}

# Group 3

# a = all stratified vs unstratified counterpart comparisons, stratified methods only
# b = all SBS bs unstratified comparisons
# c = methods better than or equal to comparisons
# abc = stratified methods that outperform their unstratified counterparts and SBS outperforms all unstratified methods

# Not in Groups 1 or 2

test.g3 <-
  test.group %>%
  group_by(SB, vnames, Method) %>%
  mutate(g3_a = (Cluster != comp_Cluster) & 
           (Sample == comp_Sample) & 
           (Cluster == "Stratified"),
         g3_b = (Method == "SBS") &
           (comp_Cluster == "Unstratified")) %>%
  filter(g3_a | g3_b) %>%
  summarise(g3_c = prod(less),
            .groups = "drop_last") %>%
  summarise(g3_abc = prod(g3_c),
            .groups = "drop") %>% 
  filter(g3_abc == 1) %>%
  anti_join(test.g1, 
            by = c("SB", "vnames")) %>%
  anti_join(test.g2, 
            by = c("SB", "vnames")) %>%
  mutate(Group = "Group 3") %>%
  select(SB, vnames, Group) %>%
  arrange(vnames, SB) 
  

test.g3

```

```{r}
figs$smd.by.scale[unique(test.g2$vnames)]

```
#### Group 4
```{r}
# Group 4

# Not in groups 1, 2 and 3

test.g4 <-
  test.group %>%
  select(SB, vnames) %>%
  unique() %>%
  anti_join(test.g1, 
            by = c("SB", "vnames")) %>%
  anti_join(test.g2, 
            by = c("SB", "vnames")) %>%
  anti_join(test.g3, 
            by = c("SB", "vnames")) %>%
  mutate(Group = "Group 4") %>%
  select(SB, vnames, Group) %>%
  arrange(vnames, SB) 

test.g4
```


```{r}
figs$smd.by.scale[unique(test.g4$vnames)]

```

#### Fine Grain Groups

```{r}
test.fine.groups <-
  bind_rows(test.g1, test.g2, test.g3, test.g4)


```
#### Fine Grain SMD plots

```{r}
tabs$smd.by.scale.fine <- 
  tabs$smd.results %>%
  left_join(test.fine.groups, 
            by = c("SB", "vnames")) %>%
  nest(smd.plot.data = -c(Variables, vnames)) %>%
  mutate(plots = pmap(.l = list(data = smd.plot.data, 
                                g.title = Variables), 
                      .f = plot_smd2, 
                      l.pos = "bottom"),
         plots = setNames(plots, vnames))

figs$smd.by.scale.fine <-
  tabs$smd.by.scale.fine$plots %>% 
  as.list

```

```{r}

t.g1 <- c("pFem", "ST.ratio")
t.g2 <- c("Urban", "Suburban", "T1")
t.g3 <- c("pELL", "dSCH", "ethHisp", "ethWhite")
t.g4 <- c("pTotfrl", "ToRu", "ethBlack", "n")

figs$smd.appx.1 <- 
  tabs$smd.results %>%
  left_join(test.fine.groups, 
            by = c("SB", "vnames")) %>%
  filter(vnames %in% t.g1) %>% 
  mutate(Gshort = str_replace(Group, "Group ", "G")) %>%
    ggplot(aes(x = RR, y = Value)) %>%
    common_line() %>%
    common_point() +
    geom_hline(yintercept = c(.25), linetype = "dotted") +
    geom_hline(yintercept = 0, linetype = "solid") + 
    scale_x_continuous(breaks = seq(10, 100, 30),
                       limits = c(10, 100)) +
    theme(legend.position = "none") +
  geom_label(aes(label = Gshort, x = 90, y = .2), size = 3) +
    facet_grid(Variables~SF_fac) 

figs$smd.appx.1 
```

```{r}


figs$smd.appx.2 <- 
  tabs$smd.results %>%
  left_join(test.fine.groups, 
            by = c("SB", "vnames")) %>%
  filter(vnames %in% t.g2) %>% 
  mutate(Gshort = str_replace(Group, "Group ", "G")) %>%
    ggplot(aes(x = RR, y = Value)) %>%
    common_line() %>%
    common_point() +
    geom_hline(yintercept = c(.25), linetype = "dotted") +
    geom_hline(yintercept = 0, linetype = "solid") + 
    scale_x_continuous(breaks = seq(10, 100, 30),
                       limits = c(10, 100)) +
    theme(legend.position = "bottom") +
  geom_label(aes(label = Gshort, x = 90, y = .75), size = 3) +
    facet_grid(Variables~SF_fac) 

figs$smd.appx.2 
```

```{r}

figs$smd.appx.3 <- 
  tabs$smd.results %>%
  left_join(test.fine.groups, 
            by = c("SB", "vnames")) %>%
  filter(vnames %in% t.g3) %>% 
  mutate(Gshort = str_replace(Group, "Group ", "G")) %>%
    ggplot(aes(x = RR, y = Value)) %>%
    common_line() %>%
    common_point() +
    geom_hline(yintercept = c(.25), linetype = "dotted") +
    geom_hline(yintercept = 0, linetype = "solid") + 
    scale_x_continuous(breaks = seq(10, 100, 30),
                       limits = c(10, 100)) +
    theme(legend.position = "bottom") +
  geom_label(aes(label = Gshort, x = 90, y = 1.25), size = 3) +
    facet_grid(Variables~SF_fac) 

figs$smd.appx.3 
```

```{r}

figs$smd.appx.4 <- 
  tabs$smd.results %>%
  left_join(test.fine.groups, 
            by = c("SB", "vnames")) %>%
  filter(vnames %in% t.g4) %>% 
  mutate(Gshort = str_replace(Group, "Group ", "G")) %>%
    ggplot(aes(x = RR, y = Value)) %>%
    common_line() %>%
    common_point() +
    geom_hline(yintercept = c(.25), linetype = "dotted") +
    geom_hline(yintercept = 0, linetype = "solid") + 
    scale_x_continuous(breaks = seq(10, 100, 30),
                       limits = c(10, 100)) +
    theme(legend.position = "bottom") +
  geom_label(aes(label = Gshort, x = 90, y = .75), size = 3) +
    facet_grid(Variables~SF_fac) 

figs$smd.appx.4 
```

### V-ratio and Log odds

```{r}

test <- 
  df.list$pop.stats %>%
  unnest(DGM) %>%
  select(vnames:log_odds, SB, Coeficient) %>%
  ungroup() %>%
  unique() %>%
  mutate(
    abs_log_odds = abs(Coeficient),
    SF_fac = paste("SF =", SB)
  ) %>%
  filter(SB > .01) %>%
  left_join(test.fine.groups, by = c("vnames", "SB"))
  
test


```

```{r}
test  %>%
  ggplot(aes(x = vratio, y = abs_log_odds, color = Group, fill = Group)) +
  # geom_mark_rect(expand = 0.04, color = "black") +
  geom_point(size = 2) +
  geom_encircle(alpha = .2, s_shape = 0, expand = .05, spread = .01, color = "black") +
  # geom_text_repel(aes(label = Variables), point.padding = .2, color = "black") +
  labs(y = lab.vars$coef,
       x = lab.vars$vratio) +
  scale_x_continuous(breaks = seq(0, 1, .1)) +
  # scale_y_continuous(breaks = seq(0, .6, .1),
  # limits = c(0, .6)) + 
  facet_wrap(~SF_fac, scales = "free_y")


```
```{r}
test  %>%
  ggplot(aes(x = vratio, y = abs_log_odds, color = Group, fill = Group)) +
  # geom_mark_rect(expand = 0.04, color = "black") +
  geom_point(size = 2) +
  geom_encircle(alpha = .2, s_shape = 0, expand = .05, spread = .01, color = "black") +
  # geom_text_repel(aes(label = Variables), point.padding = .2, color = "black") +
  labs(y = lab.vars$coef,
       x = lab.vars$vratio) +
  scale_x_continuous(breaks = seq(0, 1, .1)) +
  # scale_y_continuous(breaks = seq(0, .6, .1),
  # limits = c(0, .6)) + 
    facet_wrap(~SF_fac, scales = "free_y") +
  scale_y_log10()


```


```{r}
test %>%
  ggplot(aes(x = vratio, y = abs_log_odds, color = Group)) +
  geom_point() + 
  facet_wrap(~SF_fac, scales = "free_y")
```


```{r}
test %>%
  ggplot(aes(x = vratio, y = abs_log_odds, color = Group)) +
  geom_point() + 
  facet_wrap(~SF_fac, scales = "free_y") +
  scale_y_log10()
```


#### Test 1 - same abs_log_odds

```{r}
figs$v.coef <- 
  test %>%
  mutate(abs_log_odds = abs(log_odds)) %>%
  ggplot(aes(x = vratio, y = abs_log_odds, color = Group, fill = Group)) +
  # geom_mark_rect(expand = 0.04, color = "black") +
  geom_point(size = 2) +
  geom_encircle(alpha = .2, s_shape = 0, expand = .05, spread = .01, color = "black") +
  # geom_text_repel(aes(label = Variables), point.padding = .2, color = "black") +
  labs(y = lab.vars$coef,
       x = lab.vars$vratio) +
  scale_x_continuous(breaks = seq(0, 1, .2),
                     limits = c(0, 1)) +
  # scale_y_continuous(breaks = seq(0, .4, .1), limits = c(0, .4)) +
  scale_y_log10(n.breaks = 6,
                limits = c(.005, 1)) +
  facet_wrap(~SF_fac, scales = "free_y") +
  theme(legend.title = element_blank(), legend.position = "bottom", legend.direction = "horizontal")

figs$v.coef 
```
#### Test 2 - relative log_odds and relative vratio

```{r}
test.relative.values <- 
  df.list$pop.stats %>%
  select(vnames, vratio, log_odds) %>%
  ungroup() %>%
  mutate(
    rel_vrat = (vratio - mean(vratio)) / sd(vratio),
    rel_abs_odds = (abs(log_odds) - mean(abs(log_odds))) / sd(abs(log_odds)),
  )
```

```{r}
test %>%
  left_join(test.relative.values, 
            by = c("vnames", "vratio", "log_odds")) %>%
  ggplot(aes(x = rel_vrat, y = rel_abs_odds, color = Group, fill = Group)) +
  # ggplot(aes(x = rel_vrat, y = abs_log_odds, color = Group, fill = Group)) +
  # geom_mark_rect(expand = 0.04, color = "black") +
  geom_point(size = 2) +
  geom_encircle(alpha = .2, s_shape = 0, expand = .05, spread = .01, color = "black") +
  # geom_text_repel(aes(label = Variables), point.padding = .2, color = "black") +
  labs(y = lab.vars$coef,
       x = lab.vars$vratio) +
  # scale_x_continuous(breaks = seq(0, 1, .1)) +
  # scale_y_continuous(breaks = seq(0, .6, .1),
  # limits = c(0, .6)) + 
  facet_wrap(~SF_fac, scales = "free_y")


```
<!-- ### Test Code -->

<!-- ```{r} -->
<!-- tabs$smd.results  <-  -->
<!--   tabs$smd.results %>% -->
<!--   left_join(vars$smd.groups)  -->

<!-- figs$smd.by.group.scale <-  -->
<!--   tabs$smd.results %>% -->
<!--   ggplot(aes(x = RR,  -->
<!--              y = Value,  -->
<!--              # alpha = alpha.group, -->
<!--              color = Sample,  -->
<!--              linetype = Cluster, -->
<!--              group = interaction(Variables, Sampler))) + -->
<!--   geom_line(size = 1) + -->
<!--   annotate(geom = "rect",  -->
<!--            ymax = .25,  -->
<!--            ymin = 0,  -->
<!--            xmax = Inf,  -->
<!--            xmin = -Inf,  -->
<!--            fill = "white",  -->
<!--            alpha = .5) + -->
<!--   scale_shape_manual(values = c(1, 16)) + -->
<!--     geom_hline(yintercept = c(.25), linetype = "dotted", size = .5) + -->
<!--     geom_hline(yintercept = 0, linetype = "solid", size = .5) + -->
<!--   facet_grid(SB~Group, scales = "free_y") + -->
<!--   theme_bw() -->

<!-- figs$smd.by.group.scale  -->
<!-- ``` -->

<!-- ```{r} -->
<!-- figs$smd.by.group <- -->
<!--   tabs$smd.results  %>% -->
<!--   ggplot(aes(x = RR,  -->
<!--              y = Value,  -->
<!--              color = Sample,  -->
<!--              linetype = Cluster, -->
<!--              group = interaction(Variables, Sampler, SB))) + -->
<!--   geom_line(size = 1) + -->
<!--   annotate(geom = "rect",  -->
<!--            ymax = .25,  -->
<!--            ymin = 0,  -->
<!--            xmax = Inf,  -->
<!--            xmin = -Inf,  -->
<!--            fill = "white",  -->
<!--            alpha = .5) + -->
<!--   scale_shape_manual(values = c(1, 16)) + -->
<!--     geom_hline(yintercept = c(.25), linetype = "dotted", size = .5) + -->
<!--     geom_hline(yintercept = 0, linetype = "solid", size = .5) + -->
<!--   facet_grid(Group~Sample) + -->
<!--   theme_bw() -->

<!-- figs$smd.by.group -->

<!-- ``` -->
<!-- ```{r} -->
<!-- test <- df.list$coefs %>% -->
<!--   select(vnames, log_odds) %>% -->
<!--   right_join(tabs$smd.results) %>% -->
<!--   mutate(coef = abs(log_odds * SB)) %>% -->
<!--   select(Sampler:Sample, RR, Variables, Value, coef, log_odds, SB) -->

<!-- test %>% -->
<!--   filter(RR == 50) %>%  -->
<!--   select(-coef) %>%  -->
<!--   unique()  -->
<!-- ``` -->


<!-- ```{r} -->
<!-- tabs$smd.results %>%  -->
<!--   filter(SB == 1) %>% -->
<!--   ggplot(aes(x = RR,  -->
<!--              y = Value,  -->
<!--              color = Sample,  -->
<!--              linetype = Cluster, -->
<!--              group = interaction(Variables, Sampler, SB))) + -->
<!--   geom_line(size = 1) + -->
<!--   annotate(geom = "rect",  -->
<!--            ymax = .25,  -->
<!--            ymin = 0,  -->
<!--            xmax = Inf,  -->
<!--            xmin = -Inf,  -->
<!--            fill = "white",  -->
<!--            alpha = .5) + -->
<!--   scale_shape_manual(values = c(1, 16)) + -->
<!--     geom_hline(yintercept = c(.25), linetype = "dotted", size = .5) + -->
<!--     geom_hline(yintercept = 0, linetype = "solid", size = .5) + -->
<!--   facet_grid(Group~Sample) + -->
<!--   theme_bw() -->

<!-- ``` -->


<!-- ```{r} -->
<!-- tabs$smd.examples <-  -->
<!--   tabs$smd.results %>% -->
<!--   filter(vnames %in% vars$smd.examples)  -->

<!-- smd.examples.sep <- -->
<!--   tabs$smd.examples %>% -->
<!--   nest(data = -c(Variables, Group)) %>% -->
<!--   mutate(plots = map(data, plot_smd, l.pos = "none")) -->

<!-- figs$smd.examples <-  -->
<!--   tabs$smd.examples %>% -->
<!--   plot_smd(l.pos = "bottom") + -->
<!--   facet_grid(Variables ~ SB, scales = "free_y") -->


<!-- figs$smd.examples -->
<!-- ``` -->



<!-- ```{r} -->

<!-- df.list$smd.by.strata.results <-  -->
<!--   results$smd.stats %>% -->
<!--   # select(SB:vnames, Variables, log_odds, Sample_Catagories, mSMD, abs.mSMD) %>% -->
<!--   unpack(Sample_Catagories) %>% -->
<!--   filter(!(Rank_Factor %in% c("X9")), -->
<!--          Cluster == "Stratified", -->
<!--          strata != 0) %>% -->
<!--   select(-K, -log_odds) %>% -->
<!--   mutate(Value = abs.mSMD, -->
<!--          alpha.group = as.numeric(Value > .25)) -->

<!-- df.list$smd.by.strata.results %>% -->
<!--   filter(SB == 1, -->
<!--          vnames %in% c("ethBlack", "n")) %>% -->
<!--   plot_smd() + -->
<!--   facet_grid(Variables ~ strata) -->

<!-- ``` -->


## Feasibility

### Sampling Difficulty

```{r}
tabs$samples <- 
  results$r.stats %>% 
  unpack(Sample_Catagories) %>%
  filter(
    strata == 0,
    !(Rank_Factor %in% c("X9")),
    SB > .01
  ) %>%
  mutate(
    response_rate = accepted/contacted,
    SF_fac = paste("SF =", SB)
  )


```

#### SF ALL

```{r, fig.cap = "Schools Contacted"}
# seeds <- list()
# seeds$contacts <- floor(runif(1) * 10^4)
seeds$contacts <- 1841

tabs$samples.contacts.labels  <-
  tibble(Method_Label = c("SBS", "URS", "UCS", "SRS", "SCS"),
       RR = c(10, 10, 10, 10, 10),
       n_x = c(10, 30, 40, 10, 20),
       n_y = c(0, 200, 200, 0, 0)) %>%
  left_join(tabs$samples, 
            by = c("Method_Label", "RR")) %>%
  ungroup()

figs$samples.contacts <- 
  tabs$samples %>%
  ggplot(aes(x = RR, y = contacted)) %>%
  common_line() %>%
  common_point() +
  theme(legend.position = "none") +
  scale_x_continuous(breaks = seq(0, 100, 20),
                     limits = c(0, 100)) +
  scale_y_continuous(labels = function(x) x / 1000) +
  labs(x = lab.vars$rr,
       y = lab.vars$contacts) +
  facet_wrap(~ SB, scales = "free_y") +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$samples.contacts.labels,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = tabs$samples.contacts.labels$n_x,
                   nudge_y = tabs$samples.contacts.labels$n_y,
                   seed = seeds$contacts ,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 0)

figs$samples.contacts


```

```{r, fig.cap = "Sampling response rates"}

tabs$samples.response.rates.labels  <-
  tibble(Method_Label = c("SBS", "URS", "UCS", "SRS", "SCS"),
       RR = c(10, 20, 30, 50, 40),
       n_x = c(30, -5, -5, 30, 30),
       n_y = c(5, 20, 20, 5, 5)) %>%
  left_join(tabs$samples, 
            by = c("Method_Label", "RR")) %>%
  ungroup()  %>%
  arrange(response_rate) %>%
  ungroup() %>%
  mutate(value = response_rate * 100) 





figs$samples.response.rates <- 
  tabs$samples %>%
  mutate(value = response_rate * 100) %>%
  ggplot(aes(x = RR, y = value)) %>%
  common_line() %>%
  common_point() +
  expand_limits(y = 0) +
  theme(legend.position = "none") +
  labs(x = lab.vars$rr,
       y = lab.vars$rr.o) +
  scale_x_continuous(breaks = seq(0, 100, 20),
                     limits = c(0, 100)) +
  scale_y_continuous(breaks = seq(0, 100, 20)) +
  facet_wrap(~ SB) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$samples.response.rates.labels,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = tabs$samples.response.rates.labels$n_x,
                   nudge_y = tabs$samples.response.rates.labels$n_y,
                   seed = seeds$seeds$contacts,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 0)

figs$samples.response.rates


```

#### SF = 1
```{r, fig.cap = "Schools Contacted"}

tabs$samples.contacts.labels.sf100 <-  
  tabs$samples %>%
  filter(RR == 10,
         SB == 1) %>%
  arrange(contacted) %>%
  ungroup() %>%
  mutate(RR = 5,
         contacted = c(200, 1000, 2200, 2900, 3700))

figs$samples.contacts.sf100 <- 
  tabs$samples %>%
  filter(SB == 1) %>%
  ggplot(aes(x = RR, y = contacted)) %>%
  common_line() %>%
  common_point() +
  theme(legend.position = "none") +
  scale_x_continuous(breaks = seq(0, 100, 10)) +
  scale_y_continuous(breaks = seq(0, 4500, 500),
                     limits = c(0, 4500),
                     labels = function(x) x / 1000) +
  labs(x = lab.vars$rr,
       y = lab.vars$contacts) +
  geom_label(data = tabs$samples.contacts.labels,
             aes(label = Method_Label, color = Sample))

figs$samples.contacts.sf100


```

```{r, fig.cap = "Sampling response rates"}



tabs$samples.response.rates.labels.sf100 <-  
  tabs$samples %>%
  filter(RR == 10,
         SB == 1) %>%
  arrange(response_rate) %>%
  ungroup() %>%
  mutate(value = c(10, 25, 37, 50, 70),
         RR = 45)


figs$samples.response.rates.sf100 <- 
  tabs$samples %>%
  filter(SB == 1) %>%
  mutate(value = response_rate * 100) %>%
  ggplot(aes(x = RR, y = value)) %>%
  common_line() %>%
  common_point() +
  expand_limits(y = 0) +
  theme(legend.position = "none") +
  labs(x = lab.vars$rr,
       y = lab.vars$rr.o) +
  scale_x_continuous(breaks = seq(0, 100, 10)) +
  scale_y_continuous(breaks = seq(0, 100, 10)) +
  geom_label(data = tabs$samples.response.rates.labels,
             aes(label = Method_Label,
                 color = Sample))

figs$samples.response.rates.sf100


```
#### SF = [1, .5, .25, .01]

```{r}
figs$samples.contacts.sf.all <- 
  tabs$samples %>%
  ggplot(aes(x = RR, y = contacted)) %>%
  common_line() %>%
  common_point() +
  theme(legend.position = "none") +
  labs(x = lab.vars$rr,
       y = lab.vars$contacts) +
  scale_y_log10() +
  facet_wrap(~SF_fac)


figs$samples.contacts.sf.all

```



```{r}

figs$samples.response.rates.sf.all <- 
  tabs$samples %>%
  mutate(value = response_rate * 100) %>%
  ggplot(aes(x = RR, y = value)) %>%
  common_line() %>%
  common_point() +
  expand_limits(y = 0) +
  theme(legend.position = "none") +
  labs(x = lab.vars$rr,
       y = lab.vars$rr.o) +
  facet_wrap(~SF_fac)


figs$samples.response.rates.sf.all
```




### Difficulty Relative to Comparison Methods

#### SF = 1


```{r}
tabs$relative.counts <-
  tabs$samples %>%
  select(SB, RR, Method_Label, Cluster, Sample, contacted, accepted, response_rate) %>%
  pivot_longer(cols = c(contacted, 
                        accepted, 
                        response_rate), 
               names_to = "metric", 
               values_to = "value") %>%
  left_join(select(.,
                   SB, 
         RR, 
         metric,
         comp.method = Method_Label, 
         comp.value = value),
         by = c("SB", "RR", "metric")) %>%
  filter(Method_Label != comp.method) %>%
  mutate(relative_contacts = value / comp.value)
  
tabs$relative.counts %>%
  filter(comp.method == "URS",
         metric == "contacted")
```


```{r}

tabs$relative.counts.labs <-  
  tabs$relative.counts %>%
  filter(comp.method == "URS",
         metric == "contacted") %>%
  filter(
    SB > 0.01,
    RR == 10,
  ) %>%
  mutate(SB = paste("SF = ", SB, sep = "")) 

seeds <- list()
# seeds$fig.relative.counts <- floor(runif(1) * 10^4)
seeds$fig.relative.counts <- 1841


```

```{r}

figs$relative.counts.URS  <- 
  tabs$relative.counts %>%
  filter(comp.method == "URS",
         metric == "contacted") %>%
  mutate(SB = paste("SF = ", SB, sep = "")) %>%
  ggplot(aes(x = RR, y = relative_contacts)) %>%
  common_line() %>%
  common_point() +
  geom_hline(yintercept = 1, linetype = "dashed") +
  labs(x = lab.vars$rr,
       y = "Factor of difficulty") +
  scale_x_continuous(breaks = seq(0, 100, 20), minor_breaks = NULL, limits = c(-5, 100)) +
  scale_y_continuous(breaks = seq(0, 6, 1)) +
  facet_wrap(~ SB) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$relative.counts.labs,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = -15,
                   seed = seeds$fig.B,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 2) +
  theme_minimal() +
  theme(legend.position = "none")
  
  

figs$relative.counts.URS

```

```{r}

figs$relative.counts.UCS  <- 
  tabs$relative.counts %>%
  filter(SB == 1,
         comp.method == "UCS",
         metric == "contacted") %>%
  ggplot(aes(x = RR, y = relative_contacts)) %>%
  common_line() %>%
  common_point() +
  geom_hline(yintercept = 1, linetype = "dashed") +
  labs(x = lab.vars$rr,
       y = "Factor of difficulty") +
  scale_x_continuous(breaks = seq(0, 100, 10)) +
  scale_y_continuous(limits = c(0, 14),
                     breaks = seq(0, 14, 2))
  
  

figs$relative.counts.UCS

```

#### SF = [1, .5, .25, .01]

```{r}
# test.limits <- function(lim) {
#   rg <- abs(diff(lim)) / 10
#   
#   l.min <- max(0, min(lim) - rg)
#   l.max <- max(lim) + rg
#   
#   
#   new.lim <- 
#     c(floor(l.min * 10) / 10, 
#       ceiling(l.max * 10) / 10)
#   
#   new.lim
# }
# 
# test.breaks <- function(lim) {
#   lim = test.limits(lim)
#   
#   l.min <- min(lim)
#   l.max <- max(lim)
#   
#   new.break <- 
#     seq(l.min, 
#         l.max, 
#         (l.max - l.min) / 6)
#   
#   return(new.break)
# }
# 
# chk <- c(1.1234, 3.6543)
# 
# test.limits(chk)
# test.breaks(chk)

figs$samples.relative.contacts.sf.all <- 
  tabs$relative.counts %>%
  filter(comp.method == "URS",
         metric == "contacted") %>%
  ggplot(aes(x = RR, y = relative_contacts)) %>%
  common_line() %>%
  common_point() +
  geom_hline(yintercept = 1, linetype = "dashed") +
  labs(x = lab.vars$rr,
       y = "Factor of difficulty") +
  facet_wrap(~SB, scales = "free_y") 

figs$samples.relative.contacts.sf.all 
```
### Gini Plot

```{r}
lab.vars$gini <- "Gini Coeficient"
gini <- list()

gini$results <- results$gini


tabs$gini.coef <-
  gini$results %>%
  select(-count.data) %>%
  unpack(Sample_Catagories) %>%
  filter(
    !(Rank_Factor %in% c("X9")),
    SB > .01
  ) %>%
  group_by(RR, SB) %>%
  arrange(Gini_Coef) %>%
  mutate(
    y.pos = 1 - (1:n()/n()) / 1.5,
    x.pos = 1 - y.pos,
    SF_fac = paste("SF =", SB)
  )

# tabs$gini.coef.labels <-
#   tabs$gini.coef %>%
#   filter(RR == 10) %>%
#   mutate(RR = 5) %>%
#   arrange(Gini_Coef) %>%
#   mutate(Gini_Coef = c(.5, .65, .75, .85, .95))
```


```{r}
tabs$gini.labs <-  
  tabs$gini.coef %>%
  filter(
    !(Rank_Factor %in% c("X9")),
    SB > 0.01,
    RR == 10,
  ) 


# seeds$gini <- floor(runif(1) * 10^4)
seeds$gini <- 1841

figs$gini.coef <- tabs$gini.coef %>%
  ggplot(aes(x = RR, y = Gini_Coef)) %>%
  common_line() %>%
  common_point() +
  labs(x = lab.vars$rr,
      y = lab.vars$gini) +
  facet_grid(~ SF_fac) +
  scale_x_continuous(breaks = seq(0, 100, 20), minor_breaks = NULL, limits = c(-5, 100)) +
  scale_y_continuous(breaks = seq(0, 1, .1),
                     limits = c(0, 1), expand = c(0,0)) +
  geom_label_repel(aes(label = Method_Label, color = Sample), 
                   data = tabs$gini.labs,
                   direction = "y",
                   force_pull = 0,
                   force = .5,
                   nudge_x = -15,
                   seed = seeds$gini,
                   size = 3,
                   label.padding	= .1,
                   min.segment.length = 1) +
  theme_minimal() +
  theme(legend.position = "none")

figs$gini.coef


```

```{r}

# gini$tabs.count <-
#   gini$results %>%
#   select(SB:Sample_Catagories) %>%
#   unpack(Sample_Catagories) %>%
#   unnest(count.data)
# 
# gini$figs.count <- 
#   gini$tabs.count %>%
#   filter(SB == 1,
#          !(Rank_Factor %in% c("X9"))) %>%
#   ggplot(aes(x = c.prop.pop, y = c.prop.sel))  %>%
#   common_line() %>%
#   common_point() +
#   geom_abline(slope = 1, intercept = 0, linetype = "dotted") +
#   geom_label(aes(label = round(Gini_Coef, 2), x = x.pos, y = y.pos), 
#              data = gini$tabs.coef, size = 3) +
#   facet_wrap(~RR) +  
#   # (if (grey.plots) {scale_colour_grey(start = 0, end = .7)}) +
#   labs(x = "Cumulative proportion of population",
#          y = "Cumulative percentage of times sampled") +
#   theme_apa()
# 
# # ggsave(filename = "Figs/Presentation/GINI.jpg", dpi = 100, width = 10, height = 6)
```

# Save Data

```{r}

results$gini <- NA

```

## Size of data
```{r}
ls() %>%
  tibble(Vars = .) %>%
  mutate(MB = map(Vars, 
                  function(x) object.size(eval(parse(text = x))) %>% 
                    format(units = "MB", digits = 3))) %>%
  unnest(MB) %>%
  arrange(desc(parse_number(MB)))


tibble(
  fig = names(figs), 
  size = map_chr(figs, ~ format(object.size(.), units = "MB"))
)

tibble(
  tab = names(tabs), 
  size = map_chr(tabs, ~ format(object.size(.), units = "MB"))
)

```

## Export
```{r}
save(lab.vars,
     tabs,
     vars,
     seeds,
     file = "Data/Paper/Paper Data 02-08-2021.rdata")

save(figs,
     file = "Data/Paper/Analysis Figures 02-08-2021.rdata")

# save(gini,
#      file = "Data/Paper/Gini Data 02-08-2021.rdata")




```

# Presentation Figures

```{r, fig.cap = "Average SMDs", fig.height = 7}


# vars$plot.smd.examples <- bind_rows(
#   tibble(group = "Example", 
#          desc = "each", 
#          vnames = c("pELL", "n", "pFem")),
#   tibble(group = "ELL", 
#          desc = "ell", 
#          vnames = c("pELL")),
#   tibble(group = "N", 
#          desc = "n", 
#          vnames = c("n")),
#   tibble(group = "Fem", 
#          desc = "fem", 
#          vnames = c("pFem"))
#   )
# 
# SMDFigs <-  results$s.stats %>%
#   left_join(df.list$pop.stats) %>%
#   mutate(smd = (samp.mean - pop.mean) / pop.sd,
#          RR = parse_number(RR)) %>%
#   group_by(Sampler, RR, var, Sampling, Clustering, K) %>%
#   summarise(mSMD = mean(abs(smd)),
#             mSMD.sd = sd(abs(smd))) %>%
#   rename(vnames = var) %>%
#   left_join(df.list$coefs) %>%
#   ungroup() %>%
#   left_join(vars$plot.smd.examples ) %>%
#   group_by(group, desc) %>%
#   nest %>%
#   mutate(plots = map(data, plot_smd),
#          plots = set_names(plots, paste("smd.", desc, sep = "")),
#          data = set_names(data, paste("smd.", desc, sep = ""))) %>%
#   ungroup() %>%
#   select(data, plots)
# 
# tabs <- c(tabs, SMDFigs$data)
# figs <- c(figs, SMDFigs$plots)
# 
# figs$smd.ell
```
